requirements.txt
alabaster==0.7.12
argon2-cffi==18.3.0
arrow==0.12.1
atomicwrites==1.2.1
attrs==18.2.0
Babel==2.6.0
backcall==0.1.0
binaryornot==0.4.4
certifi==2018.10.15
cffi==1.11.5
chardet==3.0.4
Click==7.0
cookiecutter==1.6.0
coreapi==2.3.3
coreschema==0.0.4
coverage==4.5.2
decorator==4.3.0
defusedxml==0.5.0
Django==2.0.9
django-allauth==0.38.0
django-appconf==1.0.2
django-compressor==2.2
django-coverage-plugin==1.6.0
django-crispy-forms==1.7.2
django-debug-toolbar==1.10.1
django-environ==0.4.5
django-extensions==2.1.3
django-model-utils==3.1.2
django-redis==4.9.0
djangorestframework==3.9.0
docutils==0.14
factory-boy==2.11.1
Faker==1.0.0
flake8==3.6.0
future==0.17.1
idna==2.7
imagesize==1.1.0
ipdb==0.11
ipython==7.1.1
ipython-genutils==0.2.0
itypes==1.1.0
jedi==0.13.1
Jinja2==2.10
jinja2-time==0.2.0
MarkupSafe==1.1.0
mccabe==0.6.1
more-itertools==4.3.0
mypy==0.641
mypy-extensions==0.4.1
oauthlib==2.1.0
packaging==18.0
parso==0.3.1
pexpect==4.6.0
pickleshare==0.7.5
Pillow==5.3.0
pkg-resources==0.0.0
pluggy==0.8.0
poyo==0.4.2
prompt-toolkit==2.0.7
psycopg2-binary==2.7.6.1
ptyprocess==0.6.0
py==1.7.0
pycodestyle==2.4.0
pycparser==2.19
pyflakes==2.0.0
Pygments==2.2.0
pyparsing==2.3.0
pytest==4.0.0
pytest-django==3.4.4
pytest-sugar==0.9.2
python-dateutil==2.7.5
python-slugify==1.2.6
python3-openid==3.1.0
pytz==2018.7
rcssmin==1.0.6
redis==3.0.1
requests==2.20.1
requests-oauthlib==1.0.0
rjsmin==1.0.12
six==1.11.0
snowballstemmer==1.2.1
Sphinx==1.8.2
sphinxcontrib-websupport==1.1.0
sqlparse==0.2.4
termcolor==1.1.0
text-unidecode==1.2
traitlets==4.3.2
typed-ast==1.1.0
Unidecode==1.0.22
uritemplate==3.0.0
urllib3==1.24.1
wcwidth==0.1.7
Werkzeug==0.14.1
whichcraft==0.5.2
